fix(config): remove alex linter
I am removing the linting for inclusive language because the tool has not been updated
in some time. It also has an issue with ignoring rules in MDX files that is making it
difficult to work with my content. I will look for an alternative tool but it must be
removed for now.